requirements.txt
altair==4.2.2
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
config==0.5.1
db-dtypes==1.2.0
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
fonttools==4.53.0
fpdf2==2.7.9
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.64.1
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.0
oauthlib==3.2.2
openpyxl==3.1.4
packaging==24.1
pandas==2.2.2
pandas-gbq==0.23.1
pillow==10.3.0
proto-plus==1.24.0
protobuf==5.27.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydata-google-auth==1.8.2
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
six==1.16.0
smmap==5.0.1
streamlit==1.36.0
streamlit-aggrid==1.0.5
tenacity==8.4.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
